Code cleanup

+++ src/main/java/net/burningtnt/webp/vp8l/transform/PredictorTransform.java
@@ -77,7 +77,7 @@ public final class PredictorTransform implements Transform {
 
         // (0,0) Black (0x000000ff) predict
         raster.getDataElements(0, 0, rgba);
-        rgba[3] += 0xff;
+        rgba[3] += (byte) 0xff;
         raster.setDataElements(0, 0, rgba);
 
         byte[] predictor = new byte[4];
@@ -114,7 +114,7 @@ public final class PredictorTransform implements Transform {
 
                 switch (transformType) {
                     case PREDICTOR_BLACK:
-                        rgba[3] += 0xff;
+                        rgba[3] += (byte) 0xff;
                         break;
                     case PREDICTOR_L:
                         raster.getDataElements(lX, y, predictor);
